Renamed IClient And ClientSocket
