--- my-app/src/main/java/com/JavaWebServer/app/IClient.java
@@ -1,6 +1,0 @@
-package com.JavaWebServer.app;
-public interface IClient {
-  public Request getRequest();
-  public void sendResponse(String r );
-  public void close();
-}
