+++ my-app/src/main/java/com/JavaWebServer/app/InterfaceServerSocket.java
@@ -1,6 +1,6 @@
 package com.JavaWebServer.app;
 
 public interface InterfaceServerSocket {
-  public ClientSocket accept();
+  public Socket accept();
   public void close();
 }
